Update playlists.css
Seeing as how I LOVE this extension and simply can't live without it, I've taken it upon myself to attempt to address the issues it has since TTFM has last updated. This update resolves the visual UI issues and uses the latest icon sets that TTFM uses as of this commit date. Obviously, this is a first step to fixing this extension completely, and I hope Gil will accept my assistance in resolving it's issues. There is more work to be done on the JavaScript side of things, which I will attempt to address in coming updates.

Here is a more detailed summary:

- Fixed z-index of playlist and brings it to the "front layer" so you can interact with it.
- Widened the playlist by about 23px in order to fit the new icon sets.
- Fixed spacing, icons/buttons in all menus as well as z-index to bring the desired menu to the "front" for interactions. 
- Fixed song sorting (visually).
- Fixed "Move to top of queue" & "remove from playlist" buttons, as well as hover states, and updated them to use the new TTFM icons.
- Fixed start & stop Preview song icons and hover states.
- Updated playlist name section to handle really long playlist names (visually). (if that's your fancy)